Update conditions of collaborative projects.md
this was more of a philosophical question - who actually has more power in most situations is unclear. the collective power of all stakeholders is obviously greater than one startup, or even a megacorp. but those stakeholders dont have an obligation to empower the corp, so I can't stand by this statement