Update conditions of collaborative projects.md
while this is something i believe is necessary in our organization, but to require that all organizations with whom we collaborate must also follow this policy is not fair. some organizations, like the EF, pay their people crumbs, but i dont think we should refuse to work with them because of this.